✅ Extract echo_server.go
